requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-jazzmin==3.0.1
django-rest-framework==0.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
